Introducing verbatim mechanism for view rendering. AFF now renders as XML and JSON.
